created seaparate dag (airflow.py) and scrip file preprocessing.py (we have all python funtions here). Updated readme file, added old data file JPMORGAN this file we will change in our further commits
